+++ apps/docs/src/examples/select.tsx
@@ -1,13 +1,16 @@
 import { Select, Field, Button } from "@matrixzero/ui";
-
 export default function Example() {
 	return (
 		<form>
-			<Field label="地区 Region">
-				<Select name="region" defaultValue="asia">
-					<option value="asia">亚洲 Asia</option>
-					<option value="europe">欧洲 Europe</option>
-				</Select>
+			<Field label="地区 Region" required>
+				<Select
+					name="region"
+					defaultValue="asia"
+					options={[
+						{ value: "asia", label: "亚洲 Asia" },
+						{ value: "europe", label: "欧洲 Europe" },
+					]}
+				/>
 			</Field>
 			<Button type="reset">重置 Reset</Button>
 		</form>